[KernelBugTracker] Add new bridge
This adds a bridge for bugzilla.kernel.org to provide feeds for
bug comments without the need of registering an email address.

This implementation makes use of the print preview feature that
reduces bandwidth by a small margin.

Provides options to specify the number of comments to return as
well as the sorting order (latest first or oldest first)