[TwitterBridge] Optimize returned image sizes
Twitter provides an easy way to receive various image sizes based
on the same image URI:

https://dev.twitter.com/overview/api/entities-in-twitter-objects

We support different sizes: thumb, small, medium and large. The
media_url defaults to medium but you can retrieve the media in
different sizes by appending a colon + the size key (for example:
https://pbs.twimg.com/media/A7EiDWcCYAAZT1D.jpg:thumb).

-- Twitter Developer Documentation

TwitterBridge now makes use of this feature in order to provide
thumbnail images in the content and original sized images as
enclosures.

References #526